requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
colorzero==2.0
DateTime==5.5
flask==3.1.0
Flask-SocketIO==5.5.1
gpiozero==2.0.1
h11==0.14.0
importlib-metadata==4.13.0
importlib-resources==5.13.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
nrf24==2.0.0
pigpio==1.78
pyrf24==0.4.5
pyserial==3.5
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
RPi.GPIO==0.7.1
simple-websocket==1.1.0
spidev==3.6
werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
zope.interface==7.2
